+++ .gitignore
@@ -2,3 +2,4 @@ Main\config.yaml
 Main\data.json
 followers.json
 lastFollower.txt
+dist
