--- config.yaml
@@ -1,4 +1,0 @@
-API: https://api.picarto.tv/api/v1/user/followers
-Delay: '60'
-Sound: ''
-Token: ''
