--- followers.json
@@ -1,62 +1,0 @@
-[
-    {
-        "user_id": 234396,
-        "name": "Zasura",
-        "avatar": "https:\/\/images.picarto.tv\/ptvimages\/2\/23\/234396\/avatars\/4f4de3a6ae30ea2a351f1a9fcdacd031.jpg",
-        "online": false
-    },
-    {
-        "user_id": 324964,
-        "name": "MaximeJeanne",
-        "avatar": "https:\/\/images.picarto.tv\/ptvimages\/3\/32\/324964\/avatars\/78c2c59f4be99fff6017988c61d32bdd.jpg",
-        "online": false
-    },
-    {
-        "user_id": 813828,
-        "name": "Luoxenite",
-        "avatar": "https:\/\/images.picarto.tv\/ptvimages\/8\/81\/813828\/avatars\/e50ebb57433e72efd41949e634f7ad79.jpg",
-        "online": false
-    },
-    {
-        "user_id": 826265,
-        "name": "Kaykihoh",
-        "avatar": "https:\/\/images.picarto.tv\/ptvimages\/8\/82\/826265\/avatars\/dac8e0f1aa34cd9bc6981a34a5b47a31.jpg",
-        "online": false
-    },
-    {
-        "user_id": 826278,
-        "name": "MathFAKE",
-        "avatar": "https:\/\/images.picarto.tv\/ptvimages\/8\/82\/826278\/avatars\/6c074ce94d6ea1130df2d8c20d47661f.jpg",
-        "online": false
-    },
-    {
-        "user_id": 1018024,
-        "name": "TerryTigear",
-        "avatar": "https:\/\/images.picarto.tv\/ptvimages\/1\/10\/1018024\/avatars\/10137102dc9a3c55adda642a.png",
-        "online": false
-    },
-    {
-        "user_id": 1081356,
-        "name": "titof62120",
-        "avatar": "https:\/\/images.picarto.tv\/ptvimages\/1\/10\/1081356\/avatars\/aa04b2ea17989658af3a2ff1.png",
-        "online": false
-    },
-    {
-        "user_id": 1256627,
-        "name": "LexiosandCo",
-        "avatar": "https:\/\/images.picarto.tv\/ptvimages\/avatar.jpg",
-        "online": false
-    },
-    {
-        "user_id": 1261484,
-        "name": "BotSheep",
-        "avatar": "https:\/\/images.picarto.tv\/ptvimages\/avatar.jpg",
-        "online": false
-    },
-    {
-        "user_id": 1262490,
-        "name": "LexiosNco",
-        "avatar": "https:\/\/images.picarto.tv\/ptvimages\/avatar.jpg",
-        "online": false
-    }
-]
