--- lastFollower.txt
@@ -1,1 +1,0 @@
-Zasura, MaximeJeanne, Luoxenite, Kaykihoh,
